fix: use same helper on list logs
